Import HDZero OSD Driver from iNav
Mostly from https://github.com/iNavFlight/inav/pull/7668/files commit
Notable changes:
* And added missing functions to bitarray.h/.c, msp_serial.h/.c
* Used "drivers/max7456_symbols.h" instead of "drivers/osd_symbols.h"
* Used baudRates[portConfig->msp_baudrateIndex] instead of
baudRates[portConfig->peripheral_baudrateIndex]

Additional changes:
serial: serialPortFunction_e now needs 32 bits

How to use:

In CLI, Set FUNCTION_HDZERO_OSD to the appropriate serial port:
Eg. Assuming to use serial port 4, enter the following CLI commands to
enable FUNCTION_HDZERO_OSD in EmuFlight Configurator 3.0.1:
    serial 3 65536 115200 115200 0 115200
    save

+++ src/main/io/serial.c
@@ -214,11 +214,11 @@ portSharing_e determinePortSharing(const serialPortConfig_t *portConfig, serialP
     return portConfig->functionMask == function ? PORTSHARING_NOT_SHARED : PORTSHARING_SHARED;
 }
 
-bool isSerialPortShared(const serialPortConfig_t *portConfig, uint16_t functionMask, serialPortFunction_e sharedWithFunction) {
+bool isSerialPortShared(const serialPortConfig_t *portConfig, uint32_t functionMask, serialPortFunction_e sharedWithFunction) {
     return (portConfig) && (portConfig->functionMask & sharedWithFunction) && (portConfig->functionMask & functionMask);
 }
 
-serialPort_t *findSharedSerialPort(uint16_t functionMask, serialPortFunction_e sharedWithFunction) {
+serialPort_t *findSharedSerialPort(uint32_t functionMask, serialPortFunction_e sharedWithFunction) {
     for (unsigned i = 0; i < SERIAL_PORT_COUNT; i++) {
         const serialPortConfig_t *candidate = &serialConfig()->portConfigs[i];
         if (isSerialPortShared(candidate, functionMask, sharedWithFunction)) {
